Implementa ação de clique diretamente no componente
